+++ dev/src/pattern_lexer.js
@@ -3,23 +3,24 @@
     //=====================
     
     patternLexer = crossroads.patternLexer = (function(){
-
+        
         var ESCAPE_CHARS_REGEXP = /[\\.+*?\^$\[\](){}\/'#]/g, //match chars that should be escaped on string regexp
-            UNNECESSARY_SLASHES = /(?:\/(:))|(?:\/$)/g, //match slashes just before `::` and trailing slash
-
-            REQUIRED_SEGMENT_REGEXP = /([^\/]+)/, //match everything until `/`
-            OPTIONAL_SEGMENT_REGEXP = /\/?([^\/]+)?/, //match everything until `/`
+            UNNECESSARY_SLASHES_REGEXP = /\/$/g, //trailing slash
+            OPTIONAL_SLASHES_REGEXP = /([:}])\/?(:)/g, //slash between `::` or `}:`. $1 = before, $2 = after
 
             REQUIRED_PARAMS_REGEXP = /\{([^}]+)\}/g, //match everything between `{ }`
-            OPTIONAL_PARAMS_REGEXP = /:([^:]+):/g, //match everything between `:: ::`
-            PARAMS_REGEXP = /(?:\{([^}]+)\})|(?::([^:]+):)/g, //capture params name
-
-            SAVE_REQUIRED_PARAMS = '___CR_REQUIRED___', //used to save params during compile (avoid escaping unnecessary stuff)
-            SAVE_OPTIONAL_PARAMS = '___CR_OPTIONAL___',
+            OPTIONAL_PARAMS_REGEXP = /:([^:]+):/g, //match everything between `: :`
+            PARAMS_REGEXP = /(?:\{|:)([^}:]+)(?:\}|:)/g, //capture everything between `{ }` or `: :`
+            
+            //used to save params during compile (avoid escaping things that shouldn't be escaped)
+            SAVE_REQUIRED_PARAMS = '___CR_REQ___', 
+            SAVE_OPTIONAL_PARAMS = '___CR_OPT___',
+            SAVE_OPTIONAL_SLASHES = '___CR_OPT_SLASH___',
             SAVED_REQUIRED_REGEXP = new RegExp(SAVE_REQUIRED_PARAMS, 'g'),
-            SAVED_OPTIONAL_REGEXP = new RegExp(SAVE_OPTIONAL_PARAMS, 'g');
+            SAVED_OPTIONAL_REGEXP = new RegExp(SAVE_OPTIONAL_PARAMS, 'g'),
+            SAVED_OPTIONAL_SLASHES_REGEXP = new RegExp(SAVE_OPTIONAL_SLASHES, 'g');
         
-        //FIXME: not working properly for optional params because of params_regexp
+
         function getParamIds(pattern){
             var ids = [], match;
             while(match = PARAMS_REGEXP.exec(pattern)){
@@ -31,22 +32,24 @@
         function compilePattern(pattern){
             pattern = pattern || '';
             if(pattern){
-                pattern = pattern.replace(UNNECESSARY_SLASHES, '$1');
-                pattern = saveParams(pattern);
-                pattern = pattern.replace(ESCAPE_CHARS_REGEXP, '\\$&'); //make sure chars that need to be escaped are properly converted
-                pattern = convertSavedParams(pattern);
+                pattern = pattern.replace(UNNECESSARY_SLASHES_REGEXP, '');
+                pattern = tokenize(pattern);
+                pattern = pattern.replace(ESCAPE_CHARS_REGEXP, '\\$&');
+                pattern = untokenize(pattern);
             }
             return new RegExp('^'+ pattern + '/?$'); //trailing slash is optional
         }
 
-        function saveParams(pattern){
+        function tokenize(pattern){
+            pattern = pattern.replace(OPTIONAL_SLASHES_REGEXP, '$1'+ SAVE_OPTIONAL_SLASHES +'$2');
             pattern = pattern.replace(OPTIONAL_PARAMS_REGEXP, SAVE_OPTIONAL_PARAMS);
             return pattern.replace(REQUIRED_PARAMS_REGEXP, SAVE_REQUIRED_PARAMS);
         }
         
-        function convertSavedParams(pattern){
-            pattern = pattern.replace(SAVED_OPTIONAL_REGEXP, OPTIONAL_SEGMENT_REGEXP.source);
-            return pattern.replace(SAVED_REQUIRED_REGEXP, REQUIRED_SEGMENT_REGEXP.source);
+        function untokenize(pattern){
+            pattern = pattern.replace(SAVED_OPTIONAL_SLASHES_REGEXP, '\\/?');
+            pattern = pattern.replace(SAVED_OPTIONAL_REGEXP, '([^\\/]+)?\/?');
+            return pattern.replace(SAVED_REQUIRED_REGEXP, '([^\\/]+)');
         }
         
         function getParamValues(request, regexp){